fixed scheduled job

+++ src/main/java/com/piramida/cronjobs/queue/LostPendingQueueCleaner.java
@@ -3,8 +3,11 @@ package com.piramida.cronjobs.queue;
 import java.util.Calendar;
 import java.util.List;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Isolation;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -13,25 +16,38 @@ import com.piramida.dao.queue.QueueDao;
 import com.piramida.entity.PendingQueue;
 import com.piramida.entity.Queue;
 
+@Service("lostPendingQueueCleaner")
 public class LostPendingQueueCleaner {
+
+    private static Logger LOG = LoggerFactory
+	    .getLogger(LostPendingQueueCleaner.class);
 
     @Autowired
     private PendingQueueDao pendingQueueDao;
     @Autowired
     private QueueDao queueDao;
 
-    // TODO: THIS SHIT SHOULD BE TESTED!
     @Scheduled(cron = "0 0 0/2 * * ?")
     @Transactional(isolation = Isolation.SERIALIZABLE)
     public void cleanLeavedPendingQueue() {
+	if (LOG.isInfoEnabled()) {
+	    LOG.info("about to run scheduled job: delete old queues");
+	}
 	final Calendar cal = Calendar.getInstance();
 	cal.add(Calendar.HOUR_OF_DAY, -2);
 
 	final List<PendingQueue> oldQueues = pendingQueueDao
 		.findInnactiveOlderThen(cal);
-	for (final PendingQueue oldQueue : oldQueues) {
-	    final Queue garantedQueue = oldQueue.getGarantedQueue();
-	    queueDao.delete(garantedQueue);
+	System.out.println(oldQueues);
+	if (oldQueues != null) {
+	    if (LOG.isInfoEnabled()) {
+		LOG.info("about to delete {} queues", oldQueues.size());
+	    }
+	    for (final PendingQueue oldQueue : oldQueues) {
+		oldQueue.getQueue().getPendingQueues().remove(oldQueue);
+		final Queue garantedQueue = oldQueue.getGarantedQueue();
+		queueDao.delete(garantedQueue);
+	    }
 	}
 
     }
